+++ sentry-android-core/src/test/java/io/sentry/android/core/AndroidTransactionProfilerTest.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import io.sentry.CpuCollectionData
 import io.sentry.IHub
 import io.sentry.ILogger
 import io.sentry.ISentryExecutorService
@@ -363,27 +364,33 @@ class AndroidTransactionProfilerTest {
     }
 
     @Test
-    fun `profiler does not includes memory measurements when null is passed on transaction finish`() {
+    fun `profiler does not includes performance measurements when null is passed on transaction finish`() {
         val profiler = fixture.getSut(context)
         profiler.onTransactionStart(fixture.transaction1)
         val data = profiler.onTransactionFinish(fixture.transaction1, null)
         assertFalse(data!!.measurementsMap.containsKey(ProfileMeasurement.ID_MEMORY_FOOTPRINT))
         assertFalse(data.measurementsMap.containsKey(ProfileMeasurement.ID_MEMORY_NATIVE_FOOTPRINT))
+        assertFalse(data.measurementsMap.containsKey(ProfileMeasurement.ID_CPU_USAGE))
     }
 
     @Test
-    fun `profiler includes memory measurements when passed on transaction finish`() {
+    fun `profiler includes performance measurements when passed on transaction finish`() {
         val profiler = fixture.getSut(context)
         val memoryCollectionData = PerformanceCollectionData()
         memoryCollectionData.addMemoryData(MemoryCollectionData(1, 2, 3))
+        memoryCollectionData.addCpuData(CpuCollectionData(1, 1.4))
         memoryCollectionData.commitData()
         memoryCollectionData.addMemoryData(MemoryCollectionData(2, 3, 4))
         memoryCollectionData.commitData()
         profiler.onTransactionStart(fixture.transaction1)
         val data = profiler.onTransactionFinish(fixture.transaction1, memoryCollectionData)
         assertContentEquals(
+            listOf(1.4),
+            data!!.measurementsMap[ProfileMeasurement.ID_CPU_USAGE]!!.values.map { it.value }
+        )
+        assertContentEquals(
             listOf(2.0, 3.0),
-            data!!.measurementsMap[ProfileMeasurement.ID_MEMORY_FOOTPRINT]!!.values.map { it.value }
+            data.measurementsMap[ProfileMeasurement.ID_MEMORY_FOOTPRINT]!!.values.map { it.value }
         )
         assertContentEquals(
             listOf(3.0, 4.0),