+++ sentry/src/main/java/io/sentry/profilemeasurements/ProfileMeasurement.java
@@ -24,6 +24,7 @@ public final class ProfileMeasurement implements JsonUnknown, JsonSerializable {
   public static final String ID_FROZEN_FRAME_RENDERS = "frozen_frame_renders";
   public static final String ID_SLOW_FRAME_RENDERS = "slow_frame_renders";
   public static final String ID_SCREEN_FRAME_RATES = "screen_frame_rates";
+  public static final String ID_CPU_USAGE = "cpu_usage";
   public static final String ID_MEMORY_FOOTPRINT = "memory_footprint";
   public static final String ID_MEMORY_NATIVE_FOOTPRINT = "memory_native_footprint";
   public static final String ID_UNKNOWN = "unknown";
@@ -31,6 +32,7 @@ public final class ProfileMeasurement implements JsonUnknown, JsonSerializable {
   public static final String UNIT_HZ = "hz";
   public static final String UNIT_NANOSECONDS = "nanosecond";
   public static final String UNIT_BYTES = "byte";
+  public static final String UNIT_PERCENT = "percent";
   public static final String UNIT_UNKNOWN = "unknown";
 
   private @Nullable Map<String, Object> unknown;